requirements.txt
python-telegram-bot==21.10
python-dotenv==1.0.0
psutil==6.1.0
openai==1.59.5
